load, wrangles and cleans, summary
I calculated the max, min, total, average, and median played hours for all the data points, and then did the same calculations for grouped data by age, experience, and both age and experience. I'm not sure if some of these are necessary, especially since the results for grouping by both age and experience are the same.